Cleanup Redundant BlobStoreFormat Class
* No need to have an abstract class here when there's only a single impl.
